Backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
Pillow==10.4.0
pymongo==4.7.2
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
dnspython==2.6.1
numpy==1.26.4
tensorflow==2.20.0
h5py
google-genai
